# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.ApplicationResponse:
    properties:
      candidate:
        $ref: '#/definitions/dto.UserResponse'
      cover_letter:
        example: I am very excited to apply.
        type: string
      created_at:
        example: "2026-05-01T12:00:00Z"
        type: string
      cv_download_url:
        example: https://s3.amazonaws.com/...
        type: string
      id:
        example: c3d4e5f6-...
        type: string
      job:
        $ref: '#/definitions/dto.JobResponse'
      status:
        example: applied
        type: string
      updated_at:
        example: "2026-05-01T13:00:00Z"
        type: string
    type: object
  dto.ApplyJobRequest:
    properties:
      cover_letter:
        example: I am very excited to apply.
        maxLength: 2000
        type: string
      cv_key:
        example: cvs/da9d5716-4979-4f93-980a-9dfcf00cd02a/my-cv.pdf
        type: string
    type: object
  dto.AuthResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
...
  dto.CompanyResponse:
    properties:
      created_at:
        example: "2026-05-01T12:00:00Z"
        type: string
      description:
        example: A leading software company.
        type: string
      id:
        example: a1b2c3d4-...
        type: string
      industry:
        example: Technology
        type: string
      location:
        example: Colombo, Sri Lanka
        type: string
      name:
        example: TechCorp Lanka
        type: string
      website:
        example: https://techcorplanka.com
        type: string
    type: object
...
  dto.JobResponse:
    properties:
      company:
        $ref: '#/definitions/dto.CompanyResponse'
      created_at:
        example: "2026-05-01T12:00:00Z"
        type: string
      description:
        example: We are looking for an experienced Go engineer.
        type: string
      id:
        example: b2c3d4e5-...
        type: string
      job_type:
        example: full_time
        type: string
      location:
        example: Colombo, Sri Lanka
        type: string
      status:
        example: open
        type: string
      title:
        example: Senior Go Engineer
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      password:
        example: secret1234
        type: string
    required:
    - email
    - password
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: alice@example.com
        type: string
      full_name:
        example: Alice Smith
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: secret1234
        minLength: 8
        type: string
      role:
        enum:
        - candidate
        - recruiter
        example: candidate
        type: string
    required:
    - email
    - full_name
    - password
    - role
    type: object
  dto.UpdateApplicationStatusRequest:
    properties:
      status:
        enum:
        - screening
        - interview
        - offer
        - hired
        - rejected
        example: screening
        type: string
    required:
    - status
    type: object
...
  dto.UserResponse:
    properties:
      company_id:
        example: ""
        type: string
      created_at:
        example: "2026-05-01T12:00:00Z"
        type: string
      email:
        example: alice@example.com
        type: string
      full_name:
        example: Alice Smith
        type: string
      id:
        example: d4f5a2b1-...
        type: string
      role:
        example: candidate
        type: string
    type: object
  handler.APIResponse:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    email: support@hireflow.com
    name: HireFlow Support
  description: A production-grade Job Recruitment REST API built with Go and Gin.
  license:
    name: MIT
  title: HireFlow API
  version: "1.0"
paths:
  /applications/{id}:
    delete:
      description: Permanently removes the candidate's application. Candidate only.
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Withdraw an application
      tags:
      - Applications
    get:
      description: Returns a single application. Candidate sees own only; recruiter
        sees their job's only.
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ApplicationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get an application
      tags:
      - Applications
  /applications/{id}/status:
    patch:
      consumes:
      - application/json
      description: Advances or rejects an application in the pipeline. Recruiter only.
      parameters:
      - description: Application UUID
        in: path
        name: id
        required: true
        type: string
      - description: New status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateApplicationStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ApplicationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Update application status
      tags:
      - Applications
  /applications/my:
    get:
      description: Returns all applications submitted by the current candidate.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ApplicationResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: My applications
      tags:
      - Applications
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, receive a JWT token
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                errors:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Login
      tags:
      - Auth
  /auth/me:
    get:
      description: Returns the profile of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new candidate or recruiter account
      parameters:
      - description: Registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "422":
          description: Unprocessable Entity
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                errors:
                  additionalProperties:
                    type: string
                  type: object
              type: object
      summary: Register a new user
      tags:
      - Auth
...
  /jobs/{id}/applications:
    get:
      description: Lists all applications for a job. Only the recruiter who posted
        it.
      parameters:
      - description: Job UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.ApplicationResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Applications for a job
      tags:
      - Applications
  /jobs/{id}/apply:
    post:
      consumes:
      - application/json
      description: Submit an application to an open job. Candidate only.
      parameters:
      - description: Job UUID
        in: path
        name: id
        required: true
        type: string
      - description: Cover letter
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ApplyJobRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/handler.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.ApplicationResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.APIResponse'
      security:
      - BearerAuth: []
      summary: Apply to a job
      tags:
      - Applications
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your JWT token as: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"